Add CatPhan 604 and CatPhan 700 modules with corresponding classes
- Introduced CatPhan 604 phantom module (CP604) with classes for HU linearity (CTP732), image uniformity (CTP729), PSF bead MTF (CTP528CP604), and low contrast (CTP730).
- Added CatPhan 700 phantom module (CP700) with classes for HU linearity (CTP682), image uniformity (CTP712), high-resolution spatial resolution (CTP714), and low contrast (CTP515CP700).
- Created __init__.py for phantoms to facilitate module imports and maintain structure.
- Implemented backward compatibility aliases for legacy class names.
