+++ pylinac/ct.py
@@ -623,770 +623,6 @@ class CatPhanModule(Slice):
         )
 
 
-class CTP404CP504(CatPhanModule):
-    """Class for analysis of the HU linearity, geometry, and slice thickness regions of the CTP404."""
-
-    attr_name = "ctp404"
-    common_name = "HU Linearity"
-    roi_dist_mm = 58.7
-    roi_radius_mm = 5
-    roi_settings = {
-        "Air": {
-            "value": AIR,
-            "angle": -90,
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm,
-        },
-        "PMP": {
-            "value": PMP,
-            "angle": -120,
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm,
-        },
-        "LDPE": {
-            "value": LDPE,
-            "angle": 180,
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm,
-        },
-        "Poly": {
-            "value": POLY,
-            "angle": 120,
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm,
-        },
-        "Acrylic": {
-            "value": ACRYLIC,
-            "angle": 60,
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm,
-        },
-        "Delrin": {
-            "value": DELRIN,
-            "angle": 0,
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm,
-        },
-        "Teflon": {
-            "value": TEFLON,
-            "angle": -60,
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm,
-        },
-    }
-    background_roi_settings = {
-        "1": {"angle": -30, "distance": roi_dist_mm, "radius": roi_radius_mm},
-        "2": {"angle": -150, "distance": roi_dist_mm, "radius": roi_radius_mm},
-        "3": {"angle": -210, "distance": roi_dist_mm, "radius": roi_radius_mm},
-        "4": {"angle": 30, "distance": roi_dist_mm, "radius": roi_radius_mm},
-    }
-    # thickness
-    thickness_roi_height = 40
-    thickness_roi_width = 10
-    thickness_roi_distance_mm = 38
-    thickness_roi_settings = {
-        "Left": {
-            "angle": 180,
-            "width": thickness_roi_width,
-            "height": thickness_roi_height,
-            "distance": thickness_roi_distance_mm,
-        },
-        "Bottom": {
-            "angle": 90,
-            "width": thickness_roi_height,
-            "height": thickness_roi_width,
-            "distance": thickness_roi_distance_mm,
-        },
-        "Right": {
-            "angle": 0,
-            "width": thickness_roi_width,
-            "height": thickness_roi_height,
-            "distance": thickness_roi_distance_mm,
-        },
-        "Top": {
-            "angle": -90,
-            "width": thickness_roi_height,
-            "height": thickness_roi_width,
-            "distance": thickness_roi_distance_mm,
-        },
-    }
-    # geometry
-    geometry_roi_size_mm = 35
-    geometry_roi_settings = {
-        "Top-Horizontal": (0, 1),
-        "Bottom-Horizontal": (2, 3),
-        "Left-Vertical": (0, 2),
-        "Right-Vertical": (1, 3),
-    }
-    pad: str | int
-    thickness_image: Slice
-
-    def __init__(
-        self,
-        catphan,
-        offset: int,
-        hu_tolerance: float,
-        thickness_tolerance: float,
-        scaling_tolerance: float,
-        clear_borders: bool = True,
-        thickness_slice_straddle: str | int = "auto",
-        expected_hu_values: dict[str, float | int] | None = None,
-    ):
-        """
-        Parameters
-        ----------
-        catphan : `~pylinac.cbct.CatPhanBase` instance.
-        offset : int
-        hu_tolerance : float
-        thickness_tolerance : float
-        scaling_tolerance : float
-        clear_borders : bool
-        """
-        self.mm_per_pixel = catphan.mm_per_pixel
-        self.hu_tolerance = hu_tolerance
-        self.thickness_tolerance = thickness_tolerance
-        self.scaling_tolerance = scaling_tolerance
-        self.thickness_rois = {}
-        self.lines = {}
-        self.thickness_slice_straddle = thickness_slice_straddle
-        self.expected_hu_values = expected_hu_values
-        super().__init__(
-            catphan, tolerance=hu_tolerance, offset=offset, clear_borders=clear_borders
-        )
-
-    def preprocess(self, catphan) -> None:
-        # for the thickness analysis image, combine thin slices or just use one slice if slices are thick
-        if (
-            isinstance(self.thickness_slice_straddle, str)
-            and self.thickness_slice_straddle.lower() == "auto"
-        ):
-            if float(catphan.dicom_stack.metadata.SliceThickness) < 3.5:
-                self.pad = 1
-            else:
-                self.pad = 0
-        else:
-            self.pad = self.thickness_slice_straddle
-        self.thickness_image = Slice(
-            catphan,
-            combine_method="mean",
-            num_slices=self.num_slices + self.pad,
-            slice_num=self.slice_num,
-            clear_borders=self.clear_borders,
-        ).image
-
-    def _replace_hu_values(self):
-        """Possibly replace the HU values in the ROI settings with the expected values if the key is present."""
-        if self.expected_hu_values is not None:
-            for name, value in self.expected_hu_values.items():
-                if name in self.roi_settings:
-                    self.roi_settings[name]["value"] = value
-
-    def _setup_rois(self) -> None:
-        self._replace_hu_values()
-        super()._setup_rois()
-        self._setup_thickness_rois()
-        if len(self.geometry_roi_settings) > 0:
-            self._setup_geometry_rois()
-
-    def _setup_thickness_rois(self) -> None:
-        for name, setting in self.thickness_roi_settings.items():
-            self.thickness_rois[name] = ThicknessROI.from_phantom_center(
-                self.thickness_image,
-                setting["width_pixels"],
-                setting["height_pixels"],
-                setting["angle_corrected"],
-                setting["distance_pixels"],
-                self.phan_center,
-            )
-
-    def _setup_geometry_rois(self) -> None:
-        boxsize = self.geometry_roi_size_mm / self.mm_per_pixel
-        xbounds = (int(self.phan_center.x - boxsize), int(self.phan_center.x + boxsize))
-        ybounds = (int(self.phan_center.y - boxsize), int(self.phan_center.y + boxsize))
-        geo_img = self.image[ybounds[0] : ybounds[1], xbounds[0] : xbounds[1]].copy()
-        # clip to the nearest of the two extremes
-        # this can arise from direct density scans. In that case the
-        # 1 teflon node will not get detected as the edge intensity is much less than the other nodes (unlike normal)
-        # So, we clip the sub-image to the nearest extreme to the median.
-        # This does very little to normal scans. RAM-4056
-        geo_img -= np.median(geo_img)
-        nearest_extreme = min(abs(geo_img.max()), abs(geo_img.min()))
-        geo_clipped = np.clip(geo_img, a_min=-nearest_extreme, a_max=nearest_extreme)
-        geo_clipped_abs = np.abs(geo_clipped)
-        larr, regionprops, num_roi = get_regions(
-            geo_clipped_abs, fill_holes=True, clear_borders=False
-        )
-        # check that there is at least 1 ROI
-        if num_roi < 4:
-            raise ValueError("Unable to locate the Geometric nodes")
-        elif num_roi > 4:
-            regionprops = sorted(
-                regionprops, key=lambda x: x.filled_area, reverse=True
-            )[:4]
-        sorted_regions = sorted(
-            regionprops, key=lambda x: 2 * x.centroid[0] + x.centroid[1]
-        )
-        centers = [
-            Point(
-                r.weighted_centroid[1] + xbounds[0], r.weighted_centroid[0] + ybounds[0]
-            )
-            for r in sorted_regions
-        ]
-        #  setup the geometric lines
-        for name, order in self.geometry_roi_settings.items():
-            self.lines[name] = GeometricLine(
-                centers[order[0]],
-                centers[order[1]],
-                self.mm_per_pixel,
-                self.scaling_tolerance,
-            )
-
-    @property
-    def lcv(self) -> float:
-        """The low-contrast visibility"""
-        return (
-            2
-            * abs(self.rois["LDPE"].pixel_value - self.rois["Poly"].pixel_value)
-            / (self.rois["LDPE"].std + self.rois["Poly"].std)
-        )
-
-    def plotly_linearity(
-        self, plot_delta: bool = True, show_legend: bool = True
-    ) -> go.Figure:
-        fig = go.Figure()
-        nominal_x_values = [roi.nominal_val for roi in self.rois.values()]
-        if plot_delta:
-            values = [roi.value_diff for roi in self.rois.values()]
-            nominal_measurements = [0] * len(values)
-            ylabel = "HU Delta"
-        else:
-            values = [roi.pixel_value for roi in self.rois.values()]
-            nominal_measurements = nominal_x_values
-            ylabel = "Measured Values"
-        fig.add_scatter(
-            x=nominal_x_values,
-            y=values,
-            name="Measured values",
-            mode="markers",
-            marker=dict(color="green", size=10, symbol="cross", line=dict(width=1)),
-        )
-        fig.add_scatter(
-            x=nominal_x_values,
-            y=nominal_measurements,
-            mode="lines",
-            name="Nominal Values",
-            marker_color="green",
-        )
-        fig.add_scatter(
-            x=nominal_x_values,
-            y=np.array(nominal_measurements) + self.hu_tolerance,
-            mode="lines",
-            name="Upper Tolerance",
-            line=dict(dash="dash", color="red"),
-        )
-        fig.add_scatter(
-            x=nominal_x_values,
-            y=np.array(nominal_measurements) - self.hu_tolerance,
-            mode="lines",
-            name="Lower Tolerance",
-            line=dict(dash="dash", color="red"),
-        )
-        fig.update_layout(
-            xaxis_title="Nominal Values", yaxis_title=ylabel, showlegend=show_legend
-        )
-        add_title(fig, "HU Linearity")
-        return fig
-
-    def plot_linearity(
-        self, axis: plt.Axes | None = None, plot_delta: bool = True
-    ) -> tuple:
-        """Plot the HU linearity values to an axis.
-
-        Parameters
-        ----------
-        axis : None, matplotlib.Axes
-            The axis to plot the values on. If None, will create a new figure.
-        plot_delta : bool
-            Whether to plot the actual measured HU values (False), or the difference from nominal (True).
-        """
-        nominal_x_values = [roi.nominal_val for roi in self.rois.values()]
-        if axis is None:
-            fig, axis = plt.subplots()
-        if plot_delta:
-            values = [roi.value_diff for roi in self.rois.values()]
-            nominal_measurements = [0] * len(values)
-            ylabel = "HU Delta"
-        else:
-            values = [roi.pixel_value for roi in self.rois.values()]
-            nominal_measurements = nominal_x_values
-            ylabel = "Measured Values"
-        points = axis.plot(nominal_x_values, values, "g+", markersize=15, mew=2)
-        axis.plot(nominal_x_values, nominal_measurements)
-        axis.plot(
-            nominal_x_values, np.array(nominal_measurements) + self.hu_tolerance, "r--"
-        )
-        axis.plot(
-            nominal_x_values, np.array(nominal_measurements) - self.hu_tolerance, "r--"
-        )
-        axis.margins(0.05)
-        axis.grid(True)
-        axis.set_xlabel("Nominal Values")
-        axis.set_ylabel(ylabel)
-        axis.set_title("HU linearity")
-        return points
-
-    @property
-    def passed_hu(self) -> bool:
-        """Boolean specifying whether all the ROIs passed within tolerance."""
-        return all(roi.passed for roi in self.rois.values())
-
-    def plotly_rois(self, fig: go.Figure) -> None:
-        super().plotly_rois(fig)
-        # plot slice thickness / ramp ROIs
-        for name, roi in self.thickness_rois.items():
-            roi.plotly(fig, line_color="blue", name=f"Ramp {name}")
-        # plot geometry lines
-        for name, line in self.lines.items():
-            line.plotly(fig, color=line.pass_fail_color, name=f"Geometry {name}")
-
-    def plot_rois(self, axis: plt.Axes) -> None:
-        """Plot the ROIs onto the image, as well as the background ROIs"""
-        # plot HU linearity ROIs
-        super().plot_rois(axis)
-        # plot thickness ROIs
-        for roi in self.thickness_rois.values():
-            roi.plot2axes(axis, edgecolor="blue")
-        # plot geometry lines
-        for line in self.lines.values():
-            line.plot2axes(axis, color=line.pass_fail_color)
-
-    @property
-    def passed_thickness(self) -> bool:
-        """Whether the slice thickness was within tolerance from nominal."""
-        return (
-            self.slice_thickness - self.thickness_tolerance
-            < self.meas_slice_thickness
-            < self.slice_thickness + self.thickness_tolerance
-        )
-
-    @property
-    def meas_slice_thickness(self) -> float:
-        """The average slice thickness for the 4 wire measurements in mm."""
-        return np.mean(
-            sorted(
-                roi.wire_fwhm * self.mm_per_pixel * RAMP_ANGLE_RATIO
-                for roi in self.thickness_rois.values()
-            )
-        ) / (1 + 2 * self.pad)
-
-    @property
-    def avg_line_length(self) -> float:
-        return float(np.mean([line.length_mm for line in self.lines.values()]))
-
-    @property
-    def passed_geometry(self) -> bool:
-        """Returns whether all the line lengths were within tolerance."""
-        return all(line.passed for line in self.lines.values())
-
-
-class CTP404CP503(CTP404CP504):
-    """Alias for namespace consistency"""
-
-    pass
-
-
-class CTP404CP600(CTP404CP504):
-    roi_dist_mm = 58.7
-    roi_radius_mm = 5
-    roi_settings = {
-        "Air": {
-            "value": AIR,
-            "angle": 90,
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm,
-        },
-        "PMP": {
-            "value": PMP,
-            "angle": 60,
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm,
-        },
-        "LDPE": {
-            "value": LDPE,
-            "angle": 0,
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm,
-        },
-        "Poly": {
-            "value": POLY,
-            "angle": -60,
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm,
-        },
-        "Acrylic": {
-            "value": ACRYLIC,
-            "angle": -120,
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm,
-        },
-        "Delrin": {
-            "value": DELRIN,
-            "angle": -180,
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm,
-        },
-        "Teflon": {
-            "value": TEFLON,
-            "angle": 120,
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm,
-        },
-        "Vial": {
-            "value": WATER,
-            "angle": -90,
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm
-            - 1,  # the vial sits inside the ROI and needs some clearance
-        },
-    }
-
-    def _setup_rois(self) -> None:
-        """For the 600, the top ROI is an optional water vial slot. If the HU is near water we leave it, otherwise we remove it so as not to flag false failures"""
-        super()._setup_rois()
-        if self.rois["Vial"].pixel_value < -500:  # closer to air than water
-            self.rois.pop("Vial")
-
-
-class CTP404CP604(CTP404CP504):
-    roi_dist_mm = 58.7
-    roi_radius_mm = 5
-    roi_settings = {
-        "Air": {
-            "value": AIR,
-            "angle": -90,
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm,
-        },
-        "PMP": {
-            "value": PMP,
-            "angle": -120,
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm,
-        },
-        "50% Bone": {
-            "value": BONE_50,
-            "angle": -150,
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm,
-        },
-        "LDPE": {
-            "value": LDPE,
-            "angle": 180,
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm,
-        },
-        "Poly": {
-            "value": POLY,
-            "angle": 120,
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm,
-        },
-        "Acrylic": {
-            "value": ACRYLIC,
-            "angle": 60,
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm,
-        },
-        "20% Bone": {
-            "value": BONE_20,
-            "angle": 30,
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm,
-        },
-        "Delrin": {
-            "value": DELRIN,
-            "angle": 0,
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm,
-        },
-        "Teflon": {
-            "value": TEFLON,
-            "angle": -60,
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm,
-        },
-    }
-    background_roi_settings = {
-        "1": {"angle": -30, "distance": roi_dist_mm, "radius": roi_radius_mm},
-        "2": {"angle": -210, "distance": roi_dist_mm, "radius": roi_radius_mm},
-    }
-
-
-class CTP404CP700(CTP404CP504):
-    roi_dist_mm = 58.7
-    roi_radius_mm = 5
-    # The angles in the available scans are 180 - drawings.
-    # If it is found that the slices need flipping some additional logic is needed, but for now it's hardcoded
-    roi_settings = {
-        "Air": {
-            "value": AIR,
-            "angle": 180 - -90,
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm,
-        },
-        "PMP": {
-            "value": PMP,
-            "angle": 180 - -120,
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm,
-        },
-        "Lung #7112": {
-            "value": LUNG_7112,
-            "angle": 180 - -165,
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm,
-        },
-        "Delrin": {
-            "value": DELRIN,
-            "angle": 180 - 165,
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm,
-        },
-        "Poly": {
-            "value": POLY,
-            "angle": 180 - 120,
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm,
-        },
-        "Teflon": {
-            "value": TEFLON,
-            "angle": 180 - 90,
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm,
-        },
-        "Bone 20%": {
-            "value": BONE_20,
-            "angle": 180 - 60,
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm,
-        },
-        "LDPE": {
-            "value": LDPE,
-            "angle": 180 - 15,
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm,
-        },
-        "Bone 50%": {
-            "value": BONE_50,
-            "angle": 180 - -15,
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm,
-        },
-        "Acrylic": {
-            "value": ACRYLIC,
-            "angle": 180 - -60,
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm,
-        },
-        "Vial": {
-            "value": WATER,
-            "angle": 180 - -135,
-            "distance": 28,
-            "radius": roi_radius_mm,
-        },
-    }
-
-    background_roi_settings = {
-        "1": {
-            "angle": -37.5,
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm,
-        },  # NEE Between BONE_50 and ACRYLIC
-        "2": {
-            "angle": -142.5,
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm,
-        },  # NWW Between PMP and Lung_7112
-        "3": {
-            "angle": 142.5,
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm,
-        },  # SWW Between DELRIN and POLY
-        "4": {
-            "angle": 37.5,
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm,
-        },  # SEE Between BONE_20 and LDPE
-    }
-    # thickness
-    # The CatPhan 700 has wire ramps on top and bottom only.
-    # On the left and right there are bead ramps which require a different evaluation
-    thickness_roi_height = 40
-    thickness_roi_width = 10
-    thickness_roi_distance_mm = 40
-    thickness_roi_settings = {
-        "Bottom": {
-            "angle": 90,
-            "width": thickness_roi_height,
-            "height": thickness_roi_width,
-            "distance": thickness_roi_distance_mm,
-        },
-        "Top": {
-            "angle": -90,
-            "width": thickness_roi_height,
-            "height": thickness_roi_width,
-            "distance": thickness_roi_distance_mm,
-        },
-    }
-    # geometry
-    geometry_roi_settings = {}
-
-    def _setup_rois(self) -> None:
-        """For the 700, the water vial is an optional slot. If the HU is near water we leave it, otherwise we remove it so as not to flag false failures"""
-        super()._setup_rois()
-        if self.rois["Vial"].pixel_value < -500:  # closer to air than water
-            self.rois.pop("Vial")
-
-
-class CTP486(CatPhanModule):
-    """Class for analysis of the Uniformity slice of the CTP module. Measures 5 ROIs around the slice that
-    should all be close to the same value.
-    """
-
-    attr_name = "ctp486"
-    common_name = "HU Uniformity"
-    roi_dist_mm = 53
-    roi_radius_mm = 10
-    nominal_value = 0
-    nps_rois: dict[str, RectangleROI]
-    roi_settings = {
-        "Top": {
-            "value": nominal_value,
-            "angle": -90,
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm,
-        },
-        "Right": {
-            "value": nominal_value,
-            "angle": 0,
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm,
-        },
-        "Bottom": {
-            "value": nominal_value,
-            "angle": 90,
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm,
-        },
-        "Left": {
-            "value": nominal_value,
-            "angle": 180,
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm,
-        },
-        "Center": {
-            "value": nominal_value,
-            "angle": 0,
-            "distance": 0,
-            "radius": roi_radius_mm,
-        },
-    }
-
-    def plot_profiles(self, axis: plt.Axes | None = None) -> None:
-        """Plot the horizontal and vertical profiles of the Uniformity slice.
-
-        Parameters
-        ----------
-        axis : None, matplotlib.Axes
-            The axis to plot on; if None, will create a new figure.
-        """
-        if axis is None:
-            fig, axis = plt.subplots()
-        horiz_data = self.image[int(self.phan_center.y), :]
-        vert_data = self.image[:, int(self.phan_center.x)]
-        axis.plot(horiz_data, "g", label="Horizontal")
-        axis.plot(vert_data, "b", label="Vertical")
-        axis.autoscale(tight=True)
-        axis.axhline(self.nominal_value + self.tolerance, color="r", linewidth=3)
-        axis.axhline(self.nominal_value - self.tolerance, color="r", linewidth=3)
-        axis.grid(True)
-        axis.set_ylabel("HU")
-        axis.legend(loc=8, fontsize="small", title="")
-        axis.set_title("Uniformity Profiles")
-
-    def _setup_rois(self) -> None:
-        """Generate our NPS ROIs. They are just square versions of the existing ROIs."""
-        super()._setup_rois()
-        self.nps_rois = {}
-        for name, setting in self.roi_settings.items():
-            self.nps_rois[name] = RectangleROI.from_phantom_center(
-                array=self.image,
-                width=setting["radius_pixels"] * 2,
-                height=setting["radius_pixels"] * 2,
-                angle=setting["angle_corrected"],
-                dist_from_center=setting["distance_pixels"],
-                phantom_center=self.phan_center,
-            )
-
-    def plot(self, axis: plt.Axes):
-        """Plot the ROIs but also the noise power spectrum ROIs"""
-        for nps_roi in self.nps_rois.values():
-            nps_roi.plot2axes(axis, edgecolor="green", linestyle="-.")
-        super().plot(axis)
-
-    def plotly(self, **kwargs) -> go.Figure:
-        fig = super().plotly(**kwargs)
-        for name, nps_roi in self.nps_rois.items():
-            nps_roi.plotly(
-                fig, line_color="green", line_dash="dash", name=f"NPS {name}"
-            )
-        return fig
-
-    @property
-    def overall_passed(self) -> bool:
-        """Boolean specifying whether all the ROIs passed within tolerance."""
-        return all(roi.passed for roi in self.rois.values())
-
-    @property
-    def uniformity_index(self) -> float:
-        """The Uniformity Index. Elstrom et al equation 2. https://www.tandfonline.com/doi/pdf/10.3109/0284186X.2011.590525"""
-        center = self.rois["Center"]
-        uis = [
-            100 * ((roi.pixel_value - center.pixel_value) / (center.pixel_value + 1000))
-            for roi in self.rois.values()
-        ]
-        abs_uis = np.abs(uis)
-        return uis[np.argmax(abs_uis)]
-
-    @property
-    def integral_non_uniformity(self) -> float:
-        """The Integral Non-Uniformity. Elstrom et al equation 1. https://www.tandfonline.com/doi/pdf/10.3109/0284186X.2011.590525"""
-        maxhu = max(roi.pixel_value for roi in self.rois.values())
-        minhu = min(roi.pixel_value for roi in self.rois.values())
-        return (maxhu - minhu) / (maxhu + minhu + 2000)
-
-    @cached_property
-    def power_spectrum_2d(self) -> np.ndarray:
-        """The power spectrum of the uniformity ROI."""
-        return noise_power_spectrum_2d(
-            pixel_size=self.mm_per_pixel,
-            rois=[r.pixel_array for r in self.nps_rois.values()],
-        )
-
-    @cached_property
-    def power_spectrum_1d(self) -> np.ndarray:
-        """The 1D power spectrum of the uniformity ROI."""
-        return noise_power_spectrum_1d(self.power_spectrum_2d)
-
-    @property
-    def avg_noise_power(self) -> float:
-        """The average noise power of the uniformity ROI."""
-        return average_power(self.power_spectrum_1d)
-
-    @property
-    def max_noise_power_frequency(self) -> float:
-        """The frequency of the maximum noise power. 0 means no pattern."""
-        return max_frequency(self.power_spectrum_1d)
 
 
 class CTP528(abc.ABC, CatPhanModule):
@@ -1394,527 +630,6 @@ class CTP528(abc.ABC, CatPhanModule):
 
     mtf: MTF
 
-
-class CTP528CP504(CTP528):
-    """Class for analysis of the Spatial Resolution slice of the CBCT dicom data set.
-
-    A collapsed circle profile is taken of the line-pair region. This profile is search for
-    peaks and valleys. The MTF is calculated from those peaks & valleys.
-
-    Attributes
-    ----------
-
-    radius2linepairs_mm : float
-        The radius in mm to the line pairs.
-
-    """
-
-    attr_name: str = "ctp528"
-    common_name: str = "Spatial Resolution"
-    radius2linepairs_mm = 47
-    combine_method: str = "max"
-    num_slices: int = 3
-    boundaries: tuple[float, ...] = (
-        0,
-        0.107,
-        0.173,
-        0.236,
-        0.286,
-        0.335,
-        0.387,
-        0.434,
-        0.479,
-    )
-    start_angle: float = np.pi
-    ccw: bool = True
-    roi_settings = {
-        "region 1": {
-            "start": boundaries[0],
-            "end": boundaries[1],
-            "num peaks": 2,
-            "num valleys": 1,
-            "peak spacing": 0.021,
-            "gap size (cm)": 0.5,
-            "lp/mm": 0.1,
-        },
-        "region 2": {
-            "start": boundaries[1],
-            "end": boundaries[2],
-            "num peaks": 3,
-            "num valleys": 2,
-            "peak spacing": 0.01,
-            "gap size (cm)": 0.25,
-            "lp/mm": 0.2,
-        },
-        "region 3": {
-            "start": boundaries[2],
-            "end": boundaries[3],
-            "num peaks": 4,
-            "num valleys": 3,
-            "peak spacing": 0.006,
-            "gap size (cm)": 0.167,
-            "lp/mm": 0.3,
-        },
-        "region 4": {
-            "start": boundaries[3],
-            "end": boundaries[4],
-            "num peaks": 4,
-            "num valleys": 3,
-            "peak spacing": 0.00557,
-            "gap size (cm)": 0.125,
-            "lp/mm": 0.4,
-        },
-        "region 5": {
-            "start": boundaries[4],
-            "end": boundaries[5],
-            "num peaks": 4,
-            "num valleys": 3,
-            "peak spacing": 0.004777,
-            "gap size (cm)": 0.1,
-            "lp/mm": 0.5,
-        },
-        "region 6": {
-            "start": boundaries[5],
-            "end": boundaries[6],
-            "num peaks": 5,
-            "num valleys": 4,
-            "peak spacing": 0.00398,
-            "gap size (cm)": 0.083,
-            "lp/mm": 0.6,
-        },
-        "region 7": {
-            "start": boundaries[6],
-            "end": boundaries[7],
-            "num peaks": 5,
-            "num valleys": 4,
-            "peak spacing": 0.00358,
-            "gap size (cm)": 0.071,
-            "lp/mm": 0.7,
-        },
-        "region 8": {
-            "start": boundaries[7],
-            "end": boundaries[8],
-            "num peaks": 5,
-            "num valleys": 4,
-            "peak spacing": 0.0027866,
-            "gap size (cm)": 0.063,
-            "lp/mm": 0.8,
-        },
-    }
-
-    def _setup_rois(self):
-        pass
-
-    def _convert_units_in_settings(self):
-        pass
-
-    @cached_property
-    def mtf(self) -> MTF:
-        """The Relative MTF of the line pairs, normalized to the first region.
-
-        Returns
-        -------
-        dict
-        """
-        maxs = list()
-        mins = list()
-        for key, value in self.roi_settings.items():
-            max_indices, max_values = self.circle_profile.find_peaks(
-                min_distance=value["peak spacing"],
-                max_number=value["num peaks"],
-                search_region=(value["start"], value["end"]),
-            )
-            # check that the right number of peaks were found before continuing, otherwise stop searching for regions
-            if len(max_values) != value["num peaks"]:
-                break
-            maxs.append(max_values.mean())
-            _, min_values = self.circle_profile.find_valleys(
-                min_distance=value["peak spacing"],
-                max_number=value["num valleys"],
-                search_region=(min(max_indices), max(max_indices)),
-            )
-            mins.append(min_values.mean())
-        if not maxs:
-            raise ValueError(
-                "Did not find any spatial resolution pairs to analyze. File an issue on github (https://github.com/jrkerns/pylinac/issues) if this is a valid dataset."
-            )
-
-        spacings = [roi["lp/mm"] for roi in self.roi_settings.values()]
-        mtf = MTF(lp_spacings=spacings, lp_maximums=maxs, lp_minimums=mins)
-        return mtf
-
-    @property
-    def radius2linepairs(self) -> float:
-        """Radius from the phantom center to the line-pair region, corrected for pixel spacing."""
-        return self.radius2linepairs_mm * self.scaling_factor / self.mm_per_pixel
-
-    def plotly_rois(self, fig: go.Figure) -> None:
-        self.circle_profile.plotly(fig, color="blue", plot_peaks=False)
-        fig.update_layout(
-            showlegend=False,
-        )
-
-    def plot_rois(self, axis: plt.Axes) -> None:
-        """Plot the circles where the profile was taken within."""
-        self.circle_profile.plot2axes(axis, edgecolor="blue", plot_peaks=False)
-
-    @cached_property
-    def circle_profile(self) -> CollapsedCircleProfile:
-        """Calculate the median profile of the Line Pair region.
-
-        Returns
-        -------
-        :class:`pylinac.core.profile.CollapsedCircleProfile` : A 1D profile of the Line Pair region.
-        """
-        circle_profile = CollapsedCircleProfile(
-            self.phan_center,
-            self.radius2linepairs,
-            image_array=self.image,
-            start_angle=self.start_angle + np.deg2rad(self.catphan_roll),
-            width_ratio=0.04 * self.roi_size_factor,
-            sampling_ratio=2,
-            ccw=self.ccw,
-        )
-        circle_profile.filter(0.001, kind="gaussian")
-        circle_profile.ground()
-        return circle_profile
-
-
-class CTP528CP604(CTP528):
-    """Bead PSF MTF for the CatPhan 604.
-
-    Section 2 of the 604 phantom contains a 0.18 mm tungsten carbide bead.
-    We locate the bead on the HU linearity slice, extract a background-subtracted
-    2D PSF patch, radially average to 1D, then FFT to obtain the MTF curve.
-    """
-
-    attr_name: str = "ctp528"
-    common_name: str = "Spatial Resolution"
-    combine_method: str = "mean"
-    num_slices: int = 0
-    start_angle = None
-    roi_settings: dict = {}
-
-    PATCH_RADIUS_MM: float = 3.0
-    BACKGROUND_INNER_MM: float = 3.5
-    BACKGROUND_OUTER_MM: float = 6.0
-    HU_PLUG_DIST_MM: float = 58.7
-    HU_PLUG_EXCLUSION_MM: float = 12.0
-
-    def _setup_rois(self) -> None:
-        pass
-
-    def _convert_units_in_settings(self) -> None:
-        pass
-
-    @cached_property
-    def bead_center(self) -> Point:
-        """Sub-pixel location of the tungsten carbide bead."""
-        arr = self.image.array.astype(float)
-        y_idx, x_idx = np.indices(arr.shape)
-        r_mm = (
-            np.sqrt(
-                (x_idx - self.phan_center.x) ** 2
-                + (y_idx - self.phan_center.y) ** 2
-            )
-            * self.mm_per_pixel
-        )
-        phantom_radius_mm = np.sqrt(self.catphan_size / np.pi) * self.mm_per_pixel
-
-        in_plug_annulus = (
-            np.abs(r_mm - self.HU_PLUG_DIST_MM) < self.HU_PLUG_EXCLUSION_MM
-        )
-        outside_phantom = r_mm > phantom_radius_mm * 0.95
-
-        search = arr.copy()
-        search[in_plug_annulus | outside_phantom] = -np.inf
-
-        if np.all(np.isinf(search)):
-            raise ValueError(
-                "Could not locate the bead: the entire search region was masked. "
-                "Verify the phantom model and slice orientation."
-            )
-
-        peak_y, peak_x = np.unravel_index(np.argmax(search), search.shape)
-
-        half_win = max(3, int(np.ceil(self.PATCH_RADIUS_MM / self.mm_per_pixel / 2)))
-        y0 = max(0, peak_y - half_win)
-        y1 = min(arr.shape[0], peak_y + half_win + 1)
-        x0 = max(0, peak_x - half_win)
-        x1 = min(arr.shape[1], peak_x + half_win + 1)
-        win = arr[y0:y1, x0:x1].copy()
-        win_pos = np.clip(win - win.min(), 0.0, None)
-        total = win_pos.sum()
-        if total > 0:
-            ys, xs = np.indices(win.shape)
-            cx = float((xs * win_pos).sum() / total) + x0
-            cy = float((ys * win_pos).sum() / total) + y0
-        else:
-            cx, cy = float(peak_x), float(peak_y)
-
-        return Point(x=cx, y=cy)
-
-    @cached_property
-    def _psf_data(self) -> tuple[np.ndarray, tuple[float, float]]:
-        """Background-subtracted 2D PSF patch and sub-pixel bead position within it."""
-        arr = self.image.array.astype(float)
-        cx, cy = self.bead_center.x, self.bead_center.y
-
-        patch_px = int(np.ceil(self.PATCH_RADIUS_MM / self.mm_per_pixel))
-        iy, ix = int(round(cy)), int(round(cx))
-        y0 = max(0, iy - patch_px)
-        y1 = min(arr.shape[0], iy + patch_px + 1)
-        x0 = max(0, ix - patch_px)
-        x1 = min(arr.shape[1], ix + patch_px + 1)
-
-        patch = arr[y0:y1, x0:x1].copy()
-
-        y_full, x_full = np.indices(arr.shape)
-        r_from_bead = (
-            np.sqrt((x_full - cx) ** 2 + (y_full - cy) ** 2) * self.mm_per_pixel
-        )
-        bg_mask = (r_from_bead >= self.BACKGROUND_INNER_MM) & (
-            r_from_bead <= self.BACKGROUND_OUTER_MM
-        )
-        background = float(arr[bg_mask].mean()) if bg_mask.any() else 0.0
-
-        patch -= background
-        return patch, (cy - y0, cx - x0)
-
-    @property
-    def psf_patch(self) -> np.ndarray:
-        """Background-subtracted 2D PSF patch centred on the bead."""
-        return self._psf_data[0]
-
-    @cached_property
-    def mtf(self) -> BeadMTF:
-        patch, _ = self._psf_data
-        return BeadMTF(patch, self.mm_per_pixel)
-
-    def plot_rois(self, axis: plt.Axes) -> None:
-        circle = plt.Circle(
-            (self.bead_center.x, self.bead_center.y),
-            radius=self.PATCH_RADIUS_MM / self.mm_per_pixel,
-            color="blue",
-            fill=False,
-            linewidth=1.5,
-        )
-        axis.add_patch(circle)
-
-    def plotly_rois(self, fig: go.Figure) -> None:
-        theta = np.linspace(0, 2 * np.pi, 100)
-        r_px = self.PATCH_RADIUS_MM / self.mm_per_pixel
-        fig.add_scatter(
-            x=(self.bead_center.x + r_px * np.cos(theta)).tolist(),
-            y=(self.bead_center.y + r_px * np.sin(theta)).tolist(),
-            mode="lines",
-            line=dict(color="blue"),
-            name="PSF patch",
-            showlegend=False,
-        )
-
-
-class CTP528CP600(CTP528CP504):
-    start_angle = np.pi - 0.1
-    ccw = False
-    boundaries = (0, 0.127, 0.195, 0.255, 0.304, 0.354, 0.405, 0.453, 0.496)
-    roi_settings = {
-        "region 1": {
-            "start": boundaries[0],
-            "end": boundaries[1],
-            "num peaks": 2,
-            "num valleys": 1,
-            "peak spacing": 0.021,
-            "gap size (cm)": 0.5,
-            "lp/mm": 0.1,
-        },
-        "region 2": {
-            "start": boundaries[1],
-            "end": boundaries[2],
-            "num peaks": 3,
-            "num valleys": 2,
-            "peak spacing": 0.01,
-            "gap size (cm)": 0.25,
-            "lp/mm": 0.2,
-        },
-        "region 3": {
-            "start": boundaries[2],
-            "end": boundaries[3],
-            "num peaks": 4,
-            "num valleys": 3,
-            "peak spacing": 0.006,
-            "gap size (cm)": 0.167,
-            "lp/mm": 0.3,
-        },
-        "region 4": {
-            "start": boundaries[3],
-            "end": boundaries[4],
-            "num peaks": 4,
-            "num valleys": 3,
-            "peak spacing": 0.00557,
-            "gap size (cm)": 0.125,
-            "lp/mm": 0.4,
-        },
-        "region 5": {
-            "start": boundaries[4],
-            "end": boundaries[5],
-            "num peaks": 4,
-            "num valleys": 3,
-            "peak spacing": 0.004777,
-            "gap size (cm)": 0.1,
-            "lp/mm": 0.5,
-        },
-        "region 6": {
-            "start": boundaries[5],
-            "end": boundaries[6],
-            "num peaks": 5,
-            "num valleys": 4,
-            "peak spacing": 0.00398,
-            "gap size (cm)": 0.083,
-            "lp/mm": 0.6,
-        },
-        "region 7": {
-            "start": boundaries[6],
-            "end": boundaries[7],
-            "num peaks": 5,
-            "num valleys": 4,
-            "peak spacing": 0.00358,
-            "gap size (cm)": 0.071,
-            "lp/mm": 0.7,
-        },
-        "region 8": {
-            "start": boundaries[7],
-            "end": boundaries[8],
-            "num peaks": 5,
-            "num valleys": 4,
-            "peak spacing": 0.0027866,
-            "gap size (cm)": 0.063,
-            "lp/mm": 0.8,
-        },
-    }
-
-
-class CTP528CP503(CTP528CP504):
-    start_angle = 0
-    ccw = False
-    boundaries = (0, 0.111, 0.176, 0.240, 0.289, 0.339, 0.390, 0.436, 0.481)
-
-
-class CTP528CP700(CTP528):
-    """Class for analysis of the Spatial Resolution slice of the CatPhan 700 phantom
-
-    Whereas the other modules use a collapsed circle profile, the CatPhan 700 uses a
-    set of Rectangle ROIs. This is partly out of necessity as the ROIs are not
-    radially-aligned. Because of this, we also don't inherit from other CTP528 module classes.
-    """
-
-    rois: dict[str, RectangleROI]
-    attr_name: str = "ctp528"
-    common_name: str = "Spatial Resolution"
-    combine_method: str = "max"
-    num_slices: int = 3
-    start_angle = None  # this is for compatibility with other CTP528 classes where we use a circle profile.
-    # We don't use a circle profile here but when serializing results, it's expected. We use None since it doesn't apply here.
-    roi_settings = {
-        # regions are from the top-left going clockwise on CT
-        "region 1": {
-            "lp/mm": 0.1,  # from the CatPhan 700 manual
-            "radial_distance": 50,
-            "transversal_distance": -7,
-            "rotation": -90,
-            "width": 3,
-            "height": 11,
-        },
-        "region 2": {
-            "lp/mm": 0.2,
-            "radial_distance": 50,
-            "transversal_distance": 11,
-            "rotation": -90,
-            "width": 3,
-            "height": 11,
-        },
-        "region 3": {
-            "lp/mm": 0.3,
-            "radial_distance": 50,
-            "transversal_distance": -5.5,
-            "rotation": -45,
-            "width": 3,
-            "height": 10,
-        },
-        "region 4": {
-            "lp/mm": 0.4,
-            "radial_distance": 50,
-            "transversal_distance": 9.5,
-            "rotation": -45,
-            "width": 3,
-            "height": 8.5,
-        },
-        "region 5": {
-            "lp/mm": 0.5,
-            "radial_distance": 50,
-            "transversal_distance": -9,
-            "rotation": 0,
-            "width": 3,
-            "height": 8,
-        },
-        "region 6": {
-            "lp/mm": 0.6,
-            "radial_distance": 50,
-            "transversal_distance": 2,
-            "rotation": 0,
-            "width": 3,
-            "height": 7,
-        },
-        "region 7": {
-            "lp/mm": 0.7,
-            "radial_distance": 50,
-            "transversal_distance": 12,
-            "rotation": 0,
-            "width": 3,
-            "height": 6,
-        },
-        "region 8": {
-            "lp/mm": 0.8,
-            "radial_distance": 50,
-            "transversal_distance": -10.5,
-            "rotation": 45,
-            "width": 3,
-            "height": 4,
-        },
-    }
-
-    def _setup_rois(self) -> None:
-        tform_catphan_global = EuclideanTransform(
-            rotation=np.deg2rad(self.catphan_roll),
-            translation=[self.phan_center.x, self.phan_center.y],
-        )
-        for name, setting in self.roi_settings.items():
-            # The roi's are placed in polar coordinates, i.e. rotation then intrinsic translation.
-            # That's the same as extrinsic translation and then rotation which is implemented by the + operator
-            tform_roi_catphan = EuclideanTransform(
-                translation=[
-                    setting["radial_distance_pixels"],
-                    setting["transversal_distance_pixels"],
-                ]
-            ) + EuclideanTransform(rotation=np.deg2rad(setting["rotation"]))
-            # The roi in global coordinates, is the placement of the roi wrt phantom and then the phantom wrt global.
-            # This is implemented by the + operator
-            tform_roi_global = tform_roi_catphan + tform_catphan_global
-            self.rois[name] = SpatialResolutionROI(
-                array=self.image.array,
-                width=setting["width_pixels"],
-                height=setting["height_pixels"],
-                center=tform_roi_global.translation,
-                rotation=np.rad2deg(tform_roi_global.rotation),
-            )
-
-    @cached_property
-    def mtf(self) -> MTF:
-        """The Relative MTF of the line pairs, normalized to the first region."""
-        return MTF.from_high_contrast_diskset(
-            spacings=[(r["lp/mm"]) for r in self.roi_settings.values()],
-            diskset=self.rois.values(),
-        )
 
 
 class GeometricLine(Line):
@@ -1969,254 +684,6 @@ class GeometricLine(Line):
     def length_mm(self) -> float:
         """Return the length of the line in mm."""
         return self.length * self.mm_per_pixel
-
-
-class CTP515(CatPhanModule):
-    """Class for analysis of the low contrast slice of the CTP module. Low contrast is measured by obtaining
-    the average pixel value of the contrast ROIs and comparing that value to the average background value. To obtain
-    a more "human" detection level, the contrast (which is largely the same across different-sized ROIs) is multiplied
-    by the diameter. This value is compared to the contrast threshold to decide if it can be "seen".
-    """
-
-    attr_name = "ctp515"
-    common_name = "Low Contrast"
-    num_slices = 1
-    roi_dist_mm = 50
-    roi_radius_mm = [6, 3.5, 3, 2.5, 2, 1.5]
-    roi_angles = [-87.4, -69.1, -52.7, -38.5, -25.1, -12.9]
-    roi_settings = {
-        "15": {
-            "angle": roi_angles[0],
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm[0],
-        },
-        "9": {
-            "angle": roi_angles[1],
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm[1],
-        },
-        "8": {
-            "angle": roi_angles[2],
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm[2],
-        },
-        "7": {
-            "angle": roi_angles[3],
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm[3],
-        },
-        "6": {
-            "angle": roi_angles[4],
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm[4],
-        },
-        "5": {
-            "angle": roi_angles[5],
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm[5],
-        },
-    }
-    background_roi_dist_ratio = 0.75
-    background_roi_radius_mm = 4
-    WINDOW_SIZE = 50
-
-    def __init__(
-        self,
-        catphan,
-        tolerance: float,
-        cnr_threshold: float,
-        offset: int,
-        contrast_method: str,
-        visibility_threshold: float,
-        clear_borders: bool = True,
-    ):
-        self.cnr_threshold = cnr_threshold
-        self.contrast_method = contrast_method
-        self.visibility_threshold = visibility_threshold
-        super().__init__(
-            catphan, tolerance=tolerance, offset=offset, clear_borders=clear_borders
-        )
-
-    def _setup_rois(self):
-        # create both background rois dynamically, then create the actual sample ROI as normal
-        for name, setting in self.roi_settings.items():
-            self.background_rois[name + "-outer"] = (
-                LowContrastDiskROI.from_phantom_center(
-                    self.image,
-                    setting["angle_corrected"],
-                    self.background_roi_radius_mm / self.mm_per_pixel,
-                    setting["distance_pixels"] * (2 - self.background_roi_dist_ratio),
-                    self.phan_center,
-                )
-            )
-            self.background_rois[name + "-inner"] = (
-                LowContrastDiskROI.from_phantom_center(
-                    self.image,
-                    setting["angle_corrected"],
-                    self.background_roi_radius_mm / self.mm_per_pixel,
-                    setting["distance_pixels"] * self.background_roi_dist_ratio,
-                    self.phan_center,
-                )
-            )
-            background_val = float(
-                np.mean(
-                    [
-                        self.background_rois[name + "-outer"].pixel_value,
-                        self.background_rois[name + "-inner"].pixel_value,
-                    ]
-                )
-            )
-
-            self.rois[name] = LowContrastDiskROI.from_phantom_center(
-                self.image,
-                setting["angle_corrected"],
-                setting["radius_pixels"],
-                setting["distance_pixels"],
-                self.phan_center,
-                contrast_reference=background_val,
-                cnr_threshold=self.cnr_threshold,
-                contrast_method=self.contrast_method,
-                visibility_threshold=self.visibility_threshold,
-            )
-
-    @property
-    def rois_visible(self) -> int:
-        """The number of ROIs "visible"."""
-        return sum(roi.passed_visibility for roi in self.rois.values())
-
-    @property
-    def window_min(self) -> float:
-        """Lower bound of CT window/leveling to show on the plotted image. Improves apparent contrast."""
-        return (
-            Enumerable(self.background_rois.values()).min(lambda r: r.pixel_value)
-            - self.WINDOW_SIZE
-        )
-
-    @property
-    def window_max(self) -> float:
-        """Upper bound of CT window/leveling to show on the plotted image. Improves apparent contrast"""
-        return (
-            Enumerable(self.rois.values()).max(lambda r: r.pixel_value)
-            + self.WINDOW_SIZE
-        )
-
-
-class CTP515CP600(CTP515):
-    roi_angles = [
-        -87.4 + 180,
-        -69.1 + 180,
-        -52.7 + 180,
-        -38.5 + 180,
-        -25.1 + 180,
-        -12.9 + 180,
-    ]
-    roi_dist_mm = 50
-    roi_radius_mm = [6, 3.5, 3, 2.5, 2, 1.5]
-    roi_settings = {
-        "15": {
-            "angle": roi_angles[0],
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm[0],
-        },
-        "9": {
-            "angle": roi_angles[1],
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm[1],
-        },
-        "8": {
-            "angle": roi_angles[2],
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm[2],
-        },
-        "7": {
-            "angle": roi_angles[3],
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm[3],
-        },
-        "6": {
-            "angle": roi_angles[4],
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm[4],
-        },
-        "5": {
-            "angle": roi_angles[5],
-            "distance": roi_dist_mm,
-            "radius": roi_radius_mm[5],
-        },
-    }
-
-
-class CTP515CP700(CTP515CP600):
-    """Alias for namespace consistency."""
-
-    pass
-
-
-class CTP730(CTP515):
-    """Low-contrast module for Catphan 604 section 3 (CTP730).
-
-    Contains three contrast groups (0.3%, 0.5%, 1.0%) each with 9 rod diameters
-    (2, 3, 4, 5, 6, 7, 8, 9, 15 mm). The groups are arranged ~120° apart.
-    Rods appear darker than the surrounding background material.
-
-    Angles follow pylinac image convention: 0°=right, -90°=top, +90°=bottom.
-    Values are approximate based on the CTP604 manual diagram (p.24) and may
-    need refinement if the phantom sits in a non-standard rotational orientation.
-    """
-
-    attr_name = "ctp515"
-    common_name = "Low Contrast (CTP730)"
-
-    roi_dist_mm = 55
-    roi_settings = {
-        # 1.0% contrast group — upper-right to right, centred ~2 o'clock
-        "1pct_15mm": {"angle": -70, "distance": 55, "radius": 6.0},
-        "1pct_9mm":  {"angle": -60, "distance": 55, "radius": 3.5},
-        "1pct_8mm":  {"angle": -50, "distance": 55, "radius": 3.0},
-        "1pct_7mm":  {"angle": -40, "distance": 55, "radius": 2.5},
-        "1pct_6mm":  {"angle": -30, "distance": 55, "radius": 2.0},
-        "1pct_5mm":  {"angle": -20, "distance": 55, "radius": 1.5},
-        "1pct_4mm":  {"angle": -10, "distance": 55, "radius": 1.2},
-        "1pct_3mm":  {"angle":   0, "distance": 55, "radius": 1.0},
-        "1pct_2mm":  {"angle":  10, "distance": 55, "radius": 0.8},
-        # 0.5% contrast group — upper-left to left, centred ~10 o'clock
-        "05pct_15mm": {"angle":  170, "distance": 55, "radius": 6.0},
-        "05pct_9mm":  {"angle":  180, "distance": 55, "radius": 3.5},
-        "05pct_8mm":  {"angle": -170, "distance": 55, "radius": 3.0},
-        "05pct_7mm":  {"angle": -160, "distance": 55, "radius": 2.5},
-        "05pct_6mm":  {"angle": -150, "distance": 55, "radius": 2.0},
-        "05pct_5mm":  {"angle": -140, "distance": 55, "radius": 1.5},
-        "05pct_4mm":  {"angle": -130, "distance": 55, "radius": 1.2},
-        "05pct_3mm":  {"angle": -120, "distance": 55, "radius": 1.0},
-        "05pct_2mm":  {"angle": -110, "distance": 55, "radius": 0.8},
-        # 0.3% contrast group — bottom, centred ~6 o'clock
-        "03pct_15mm": {"angle":  50, "distance": 55, "radius": 6.0},
-        "03pct_9mm":  {"angle":  60, "distance": 55, "radius": 3.5},
-        "03pct_8mm":  {"angle":  70, "distance": 55, "radius": 3.0},
-        "03pct_7mm":  {"angle":  80, "distance": 55, "radius": 2.5},
-        "03pct_6mm":  {"angle":  90, "distance": 55, "radius": 2.0},
-        "03pct_5mm":  {"angle": 100, "distance": 55, "radius": 1.5},
-        "03pct_4mm":  {"angle": 110, "distance": 55, "radius": 1.2},
-        "03pct_3mm":  {"angle": 120, "distance": 55, "radius": 1.0},
-        "03pct_2mm":  {"angle": 130, "distance": 55, "radius": 0.8},
-    }
-    background_roi_dist_ratio = 0.75
-    background_roi_radius_mm = 4
-
-    @property
-    def rois_1pct(self) -> dict:
-        """ROIs belonging to the 1.0% contrast group."""
-        return {k: v for k, v in self.rois.items() if k.startswith("1pct_")}
-
-    @property
-    def rois_05pct(self) -> dict:
-        """ROIs belonging to the 0.5% contrast group."""
-        return {k: v for k, v in self.rois.items() if k.startswith("05pct_")}
-
-    @property
-    def rois_03pct(self) -> dict:
-        """ROIs belonging to the 0.3% contrast group."""
-        return {k: v for k, v in self.rois.items() if k.startswith("03pct_")}
 
 
 class CatPhanBase(ResultsDataMixin[CatphanResult], QuaacMixin):
@@ -3255,257 +1722,6 @@ class CatPhanBase(ResultsDataMixin[CatphanResult], QuaacMixin):
         return data
 
 
-@capture_warnings
-class CatPhan503(CatPhanBase):
-    """A class for loading and analyzing CT DICOM files of a CatPhan 503.
-    Analyzes: Uniformity (CTP486), High-Contrast Spatial Resolution (CTP528), Image Scaling & HU Linearity (CTP404).
-    """
-
-    _demo_url = "CatPhan503.zip"
-    _model = "503"
-    catphan_radius_mm = 97
-    modules = {
-        CTP404CP503: {"offset": 0},
-        CTP486: {"offset": -110},
-        CTP528CP503: {"offset": -30},
-    }
-
-    @staticmethod
-    def run_demo(show: bool = True):
-        """Run the CBCT demo using high-quality head protocol images."""
-        cbct = CatPhan503.from_demo_images()
-        cbct.analyze()
-        print(cbct.results())
-        cbct.plot_analyzed_image(show)
-
-
-@capture_warnings
-class CatPhan504(CatPhanBase):
-    """A class for loading and analyzing CT DICOM files of a CatPhan 504. Can be from a CBCT or CT scanner
-    Analyzes: Uniformity (CTP486), High-Contrast Spatial Resolution (CTP528),
-    Image Scaling & HU Linearity (CTP404), and Low contrast (CTP515).
-    """
-
-    _demo_url = "CatPhan504.zip"
-    _model = "504"
-    catphan_radius_mm = 101
-    modules = {
-        CTP404CP504: {"offset": 0},
-        CTP486: {"offset": -65},
-        CTP528CP504: {"offset": 30},
-        CTP515: {"offset": -30},
-    }
-
-    @staticmethod
-    def run_demo(show: bool = True):
-        """Run the CBCT demo using high-quality head protocol images."""
-        cbct = CatPhan504.from_demo_images()
-        cbct.analyze()
-        print(cbct.results())
-        cbct.plot_analyzed_image(show)
-
-
-@capture_warnings
-class CatPhan604(CatPhanBase):
-    """A class for loading and analyzing CT DICOM files of a CatPhan 604. Can be from a CBCT or CT scanner
-    Analyzes: Uniformity (CTP486), High-Contrast Spatial Resolution (CTP528),
-    Image Scaling & HU Linearity (CTP404), and Low contrast (CTP515).
-    """
-
-    _demo_url = "CatPhan604.zip"
-    _model = "604"
-    catphan_radius_mm = 101
-    modules = {
-        CTP404CP604: {"offset": 0},
-        CTP486: {"offset": -80},
-        CTP528CP604: {"offset": 0},  # bead is on Section 2, same z as CTP404
-        CTP730: {"offset": -40},  # CTP730 low-contrast section at +40 mm from Section 2
-    }
-
-    @staticmethod
-    def run_demo(show: bool = True):
-        """Run the CBCT demo using high-quality head protocol images."""
-        cbct = CatPhan604.from_demo_images()
-        cbct.analyze()
-        print(cbct.results())
-        cbct.plot_analyzed_image(show)
-
-    def refine_origin_slice(self, initial_slice_num: int) -> int:
-        """The HU plugs are longer than the 'wire section'. This applies a refinement to find the
-        slice that has the least angle between the centers of the left and right wires.
-
-        Under normal conditions, we would simply apply an offset to the
-        initial slice. The rods extend 4-5mm past the wire section.
-        Unfortunately, we also sometimes need to account for users with the
-        RM R1-4 jig which will cause localization issues due to the base of the
-        jig touching the phantom.
-        This solution is robust to the jig being present, but can suffer
-        from images where the angle of the wire, due to noise,
-        appears small but doesn't actually represent the wire ramp.
-
-        Starting with the initial slice, go +/- 5 slices to find the slice with the least angle
-        between the left and right wires.
-
-        This suffers from a weakness that the roll is not yet determined.
-        This will thus return the slice that has the least ABSOLUTE
-        roll. If the phantom has an inherent roll, this will not be detected and may be off by a slice or so.
-        Given the angle of the wire, the error would be small and likely only 1-2 slices max.
-        """
-        angles = []
-        # make a CTP module for the purpose of easily extracting the thickness ROIs
-        ctp404, offset = self._get_module(CTP404CP604, raise_empty=True)
-        # we don't want to set up our other ROIs (they sometimes fail) so we temporarily override the method
-        original_setup = ctp404._setup_rois
-        ctp404._setup_rois = lambda x: x
-        ctp = ctp404(
-            self,
-            offset=offset,
-            clear_borders=self.clear_borders,
-            hu_tolerance=0,
-            scaling_tolerance=0,
-            thickness_tolerance=0,
-        )
-        # we have to reset the method after we're done for future calls
-        ctp404._setup_rois = original_setup
-        for slice_num in range(initial_slice_num - 5, initial_slice_num + 5):
-            slice = Slice(self, slice_num, clear_borders=self.clear_borders)
-            # make a slice and add the wire ROIs to it.
-            troi = {}
-            for name, setting in ctp.thickness_roi_settings.items():
-                troi[name] = ThicknessROI.from_phantom_center(
-                    slice.image,
-                    setting["width_pixels"],
-                    setting["height_pixels"],
-                    setting["angle_corrected"],
-                    setting["distance_pixels"],
-                    slice.phan_center,
-                )
-            # now find the angle between the left and right and top and bottom wires via the long profile
-            left_wire = troi["Left"].long_profile.center_idx
-            right_wire = troi["Right"].long_profile.center_idx
-            h_angle = abs(left_wire - right_wire)
-            top_wire = troi["Top"].long_profile.center_idx
-            bottom_wire = troi["Bottom"].long_profile.center_idx
-            v_angle = abs(top_wire - bottom_wire)
-            angle = (h_angle + v_angle) / 2
-
-            angles.append(
-                {
-                    "slice": slice_num,
-                    "angle": angle,
-                    "left width": troi["Left"].long_profile.field_width_px,
-                    "right width": troi["Right"].long_profile.field_width_px,
-                    # the values AT the FWXM
-                    "left center": troi["Left"].long_profile.y_at_x(left_wire),
-                    "right center": troi["Right"].long_profile.y_at_x(right_wire),
-                    "left profile": troi["Left"].long_profile.values,
-                    "right profile": troi["Right"].long_profile.values,
-                }
-            )
-
-        # some slices might not include the wire
-        # we need to drop those; we do so by dropping pairs that have a field width well below the median
-        # or by fields who don't appear to have the wire in them (max value is near median)
-        median_width_l = np.median([angle["left width"] for angle in angles])
-        median_width_r = np.median([angle["right width"] for angle in angles])
-        median_width = (median_width_l + median_width_r) / 2
-        # get median and max pixel values of all the profiles
-        median_left_pixel_val = np.median(
-            np.concatenate([a["left profile"] for a in angles])
-        )
-        median_right_pixel_val = np.median(
-            np.concatenate([a["right profile"] for a in angles])
-        )
-        median_pixel_val = (median_left_pixel_val + median_right_pixel_val) / 2
-        max_left_pixel_val = np.max(np.concatenate([a["left profile"] for a in angles]))
-        max_right_pixel_val = np.max(
-            np.concatenate([a["right profile"] for a in angles])
-        )
-        max_pixel_val = (max_left_pixel_val + max_right_pixel_val) / 2
-
-        for angle_set in angles.copy():
-            # field width is well below the median; probably not in the slice; drop it
-            if (
-                angle_set["left width"] < median_width * 0.7
-                or angle_set["right width"] < median_width * 0.7
-            ):
-                angles.remove(angle_set)
-                continue
-            # if the max pixel value of the angle set is closer to the overall median than the max
-            # it means the wire isn't in the slice; drop it
-            fwxm_pixel = np.mean((angle_set["left center"], angle_set["right center"]))
-            delta_median = abs(median_pixel_val - fwxm_pixel)
-            delta_max = abs(max_pixel_val - fwxm_pixel)
-            if delta_median < delta_max:
-                angles.remove(angle_set)
-
-        # now find the slice with the least angle, accounting for the phantom roll
-        m_slice_num = np.argsort([a["angle"] - self.catphan_roll for a in angles])
-        refined_slice = angles[m_slice_num[0]]["slice"]
-        return refined_slice
-
-
-@capture_warnings
-class CatPhan600(CatPhanBase):
-    """A class for loading and analyzing CT DICOM files of a CatPhan 600.
-    Analyzes: Uniformity (CTP486), High-Contrast Spatial Resolution (CTP528),
-    Image Scaling & HU Linearity (CTP404), and Low contrast (CTP515).
-    """
-
-    _demo_url = "CatPhan600.zip"
-    _model = "600"
-    catphan_radius_mm = 101
-    modules = {
-        CTP404CP600: {"offset": 0},
-        CTP486: {"offset": -160},
-        CTP515CP600: {"offset": -110},
-        CTP528CP600: {"offset": -70},
-    }
-
-    @staticmethod
-    def run_demo(show: bool = True):
-        """Run the CatPhan 600 demo."""
-        cbct = CatPhan600.from_demo_images()
-        cbct.analyze()
-        print(cbct.results())
-        cbct.plot_analyzed_image(show)
-
-    def find_phantom_roll(self, func: Callable | None = None) -> float:
-        """With the CatPhan 600, we have to consider that the top air ROI
-        has a water vial in it (see pg 12 of the manual). If so, the top air ROI won't be detected.
-        Rather, the default algorithm will find the bottom air ROI and teflon to the left.
-        It may also find the top air ROI if the water vial isn't there.
-        We use the below lambda to select the bottom air and teflon ROIs consistently.
-        These two ROIs are at 75 degrees from cardinal. We thus offset the default outcome by 75.
-
-        HOWEVER, for direct density scans, the Teflon ROI might not register because of the reduced
-        HU. 🤦‍♂️. We make a best guess depending on the detected roll. If it's ~75 degrees,
-        we have caught the bottom Air and Teflon. If it's near zero, we have
-        caught the top and bottom Air ROIs.
-        """
-        angle = super().find_phantom_roll(lambda x: -x.centroid[0])
-        if abs(angle) < 10:
-            return angle
-        return angle + 75
-
-
-@capture_warnings
-class CatPhan700(CatPhanBase):
-    """A class for loading and analyzing CT DICOM files of a CatPhan 700.
-    Analyzes: Uniformity (CTP486/CTP712), High-Contrast Spatial Resolution (CTP714),
-    Image Scaling & HU Linearity (CTP404CP700/CTP682), Low contrast (CTP515) and slice geometry (CTP721)
-    """
-
-    # _demo_url = "CatPhan600.zip"
-    _model = "700"
-    catphan_radius_mm = 101
-    modules = {
-        CTP404CP700: {"offset": 0},
-        CTP515CP700: {"offset": -80},
-        CTP486: {"offset": -160},
-        CTP528CP700: {"offset": -40},
-    }
-
 
 def get_regions(
     slice_or_arr: Slice | np.array,
@@ -3594,3 +1810,49 @@ def rois_to_results(dict_mapping: dict[str, DiskROI]) -> dict[str, ROIResult]:
             passed=getattr(roi, "passed", None),
         )
     return flat_dict
+
+# ---------------------------------------------------------------------------
+# Re-exports from phantom sub-modules
+#
+# The phantom-specific classes have been moved to pylinac/phantoms/.
+# These re-exports restore all original ct.py names so existing user code
+# keeps working without change.  New code should import from the phantom
+# sub-modules directly (e.g. ``from pylinac.phantoms.CP504 import CatPhan504``).
+# ---------------------------------------------------------------------------
+
+from .phantoms.CP504 import (
+    CTP404 as CTP404CP504,
+    CTP404CP503,
+    CTP486,
+    CTP515,
+    CTP528CP504,
+    CTP528CP503,
+    CatPhan503,
+    CatPhan504,
+)
+from .phantoms.CP600 import (
+    CTP763,
+    CTP764,
+    CTP515CP600,
+    CTP591,
+    CatPhan600,
+    CTP404CP600,
+    CTP528CP600,
+)
+from .phantoms.CP604 import (
+    CTP732,
+    CTP729,
+    CTP528CP604,
+    CTP730,
+    CatPhan604,
+    CTP404CP604,
+)
+from .phantoms.CP700 import (
+    CTP682,
+    CTP712,
+    CTP714,
+    CTP515CP700,
+    CatPhan700,
+    CTP404CP700,
+    CTP528CP700,
+)
