+++ pylinac/phantoms/CP604.py
@@ -1,0 +1,464 @@
+"""CatPhan 604 phantom module classes.
+
+Physical modules:
+  CTP732 — HU linearity (adds 50%/20% Bone inserts vs CTP404)
+  CTP729 — Image uniformity (identical to CTP486)
+  CTP528CP604 — PSF bead MTF (replaces line-pair gauge)
+  CTP730 — Low contrast (27 rods, 3 contrast groups)
+"""
+
+from __future__ import annotations
+
+from functools import cached_property
+
+import matplotlib.pyplot as plt
+import numpy as np
+from plotly import graph_objects as go
+
+from ..core.mtf import BeadMTF
+from ..core.warnings import capture_warnings
+from ..ct import (
+    BONE_20,
+    BONE_50,
+    AIR,
+    ACRYLIC,
+    DELRIN,
+    LDPE,
+    POLY,
+    PMP,
+    TEFLON,
+    CatPhanBase,
+    CTP528 as _CTP528Base,
+    Point,
+    Slice,
+    ThicknessROI,
+)
+from . import CP504
+
+
+# ---------------------------------------------------------------------------
+# CTP732 — HU linearity (CatPhan 604)
+# ---------------------------------------------------------------------------
+
+class CTP732(CP504.CTP404):
+    """HU linearity module for CatPhan 604 (CTP732).
+
+    Adds 50% Bone (at −150°) and 20% Bone (at +30°) inserts compared with the
+    standard CTP404.  Only two background ROIs instead of four.
+    """
+
+    roi_dist_mm = 58.7
+    roi_radius_mm = 5
+    roi_settings = {
+        "Air": {
+            "value": AIR,
+            "angle": -90,
+            "distance": roi_dist_mm,
+            "radius": roi_radius_mm,
+        },
+        "PMP": {
+            "value": PMP,
+            "angle": -120,
+            "distance": roi_dist_mm,
+            "radius": roi_radius_mm,
+        },
+        "50% Bone": {
+            "value": BONE_50,
+            "angle": -150,
+            "distance": roi_dist_mm,
+            "radius": roi_radius_mm,
+        },
+        "LDPE": {
+            "value": LDPE,
+            "angle": 180,
+            "distance": roi_dist_mm,
+            "radius": roi_radius_mm,
+        },
+        "Poly": {
+            "value": POLY,
+            "angle": 120,
+            "distance": roi_dist_mm,
+            "radius": roi_radius_mm,
+        },
+        "Acrylic": {
+            "value": ACRYLIC,
+            "angle": 60,
+            "distance": roi_dist_mm,
+            "radius": roi_radius_mm,
+        },
+        "20% Bone": {
+            "value": BONE_20,
+            "angle": 30,
+            "distance": roi_dist_mm,
+            "radius": roi_radius_mm,
+        },
+        "Delrin": {
+            "value": DELRIN,
+            "angle": 0,
+            "distance": roi_dist_mm,
+            "radius": roi_radius_mm,
+        },
+        "Teflon": {
+            "value": TEFLON,
+            "angle": -60,
+            "distance": roi_dist_mm,
+            "radius": roi_radius_mm,
+        },
+    }
+    background_roi_settings = {
+        "1": {"angle": -30, "distance": roi_dist_mm, "radius": roi_radius_mm},
+        "2": {"angle": -210, "distance": roi_dist_mm, "radius": roi_radius_mm},
+    }
+
+
+# ---------------------------------------------------------------------------
+# CTP729 — Image uniformity (identical hardware to CTP486)
+# ---------------------------------------------------------------------------
+
+class CTP729(CP504.CTP486):
+    """Image uniformity module for CatPhan 604 (CTP729).
+
+    Functionally and geometrically identical to CTP486.
+    """
+
+    pass
+
+
+# ---------------------------------------------------------------------------
+# CTP528CP604 — PSF bead MTF
+# ---------------------------------------------------------------------------
+
+class CTP528CP604(_CTP528Base):
+    """Bead PSF MTF for the CatPhan 604.
+
+    Section 2 of the 604 phantom contains a 0.18 mm tungsten carbide bead.
+    We locate the bead on the HU linearity slice, extract a background-subtracted
+    2D PSF patch, radially average to 1D, then FFT to obtain the MTF curve.
+    """
+
+    attr_name: str = "ctp528"
+    common_name: str = "Spatial Resolution"
+    combine_method: str = "mean"
+    num_slices: int = 0
+    start_angle = None
+    roi_settings: dict = {}
+
+    PATCH_RADIUS_MM: float = 3.0
+    BACKGROUND_INNER_MM: float = 3.5
+    BACKGROUND_OUTER_MM: float = 6.0
+    HU_PLUG_DIST_MM: float = 58.7
+    HU_PLUG_EXCLUSION_MM: float = 12.0
+
+    def _setup_rois(self) -> None:
+        pass
+
+    def _convert_units_in_settings(self) -> None:
+        pass
+
+    @cached_property
+    def bead_center(self) -> Point:
+        """Sub-pixel location of the tungsten carbide bead."""
+        arr = self.image.array.astype(float)
+        y_idx, x_idx = np.indices(arr.shape)
+        r_mm = (
+            np.sqrt(
+                (x_idx - self.phan_center.x) ** 2
+                + (y_idx - self.phan_center.y) ** 2
+            )
+            * self.mm_per_pixel
+        )
+        phantom_radius_mm = np.sqrt(self.catphan_size / np.pi) * self.mm_per_pixel
+
+        in_plug_annulus = (
+            np.abs(r_mm - self.HU_PLUG_DIST_MM) < self.HU_PLUG_EXCLUSION_MM
+        )
+        outside_phantom = r_mm > phantom_radius_mm * 0.95
+
+        search = arr.copy()
+        search[in_plug_annulus | outside_phantom] = -np.inf
+
+        if np.all(np.isinf(search)):
+            raise ValueError(
+                "Could not locate the bead: the entire search region was masked. "
+                "Verify the phantom model and slice orientation."
+            )
+
+        peak_y, peak_x = np.unravel_index(np.argmax(search), search.shape)
+
+        half_win = max(3, int(np.ceil(self.PATCH_RADIUS_MM / self.mm_per_pixel / 2)))
+        y0 = max(0, peak_y - half_win)
+        y1 = min(arr.shape[0], peak_y + half_win + 1)
+        x0 = max(0, peak_x - half_win)
+        x1 = min(arr.shape[1], peak_x + half_win + 1)
+        win = arr[y0:y1, x0:x1].copy()
+        win_pos = np.clip(win - win.min(), 0.0, None)
+        total = win_pos.sum()
+        if total > 0:
+            ys, xs = np.indices(win.shape)
+            cx = float((xs * win_pos).sum() / total) + x0
+            cy = float((ys * win_pos).sum() / total) + y0
+        else:
+            cx, cy = float(peak_x), float(peak_y)
+
+        return Point(x=cx, y=cy)
+
+    @cached_property
+    def _psf_data(self) -> tuple[np.ndarray, tuple[float, float]]:
+        """Background-subtracted 2D PSF patch and sub-pixel bead position within it."""
+        arr = self.image.array.astype(float)
+        cx, cy = self.bead_center.x, self.bead_center.y
+
+        patch_px = int(np.ceil(self.PATCH_RADIUS_MM / self.mm_per_pixel))
+        iy, ix = int(round(cy)), int(round(cx))
+        y0 = max(0, iy - patch_px)
+        y1 = min(arr.shape[0], iy + patch_px + 1)
+        x0 = max(0, ix - patch_px)
+        x1 = min(arr.shape[1], ix + patch_px + 1)
+
+        patch = arr[y0:y1, x0:x1].copy()
+
+        y_full, x_full = np.indices(arr.shape)
+        r_from_bead = (
+            np.sqrt((x_full - cx) ** 2 + (y_full - cy) ** 2) * self.mm_per_pixel
+        )
+        bg_mask = (r_from_bead >= self.BACKGROUND_INNER_MM) & (
+            r_from_bead <= self.BACKGROUND_OUTER_MM
+        )
+        background = float(arr[bg_mask].mean()) if bg_mask.any() else 0.0
+
+        patch -= background
+        return patch, (cy - y0, cx - x0)
+
+    @property
+    def psf_patch(self) -> np.ndarray:
+        """Background-subtracted 2D PSF patch centred on the bead."""
+        return self._psf_data[0]
+
+    @cached_property
+    def mtf(self) -> BeadMTF:
+        patch, _ = self._psf_data
+        return BeadMTF(patch, self.mm_per_pixel)
+
+    def plot_rois(self, axis: plt.Axes) -> None:
+        circle = plt.Circle(
+            (self.bead_center.x, self.bead_center.y),
+            radius=self.PATCH_RADIUS_MM / self.mm_per_pixel,
+            color="blue",
+            fill=False,
+            linewidth=1.5,
+        )
+        axis.add_patch(circle)
+
+    def plotly_rois(self, fig: go.Figure) -> None:
+        theta = np.linspace(0, 2 * np.pi, 100)
+        r_px = self.PATCH_RADIUS_MM / self.mm_per_pixel
+        fig.add_scatter(
+            x=(self.bead_center.x + r_px * np.cos(theta)).tolist(),
+            y=(self.bead_center.y + r_px * np.sin(theta)).tolist(),
+            mode="lines",
+            line=dict(color="blue"),
+            name="PSF patch",
+            showlegend=False,
+        )
+
+
+# ---------------------------------------------------------------------------
+# CTP730 — Low contrast (27 rods, 3 contrast groups)
+# ---------------------------------------------------------------------------
+
+class CTP730(CP504.CTP515):
+    """Low-contrast module for CatPhan 604 (CTP730).
+
+    Contains three contrast groups (0.3 %, 0.5 %, 1.0 %) each with 9 rod
+    diameters (2, 3, 4, 5, 6, 7, 8, 9, 15 mm).  The groups are arranged
+    ~120° apart.  Rods appear darker than the surrounding background material.
+
+    Angles follow pylinac image convention: 0°=right, −90°=top, +90°=bottom.
+    Values are approximate based on the CTP604 manual diagram (p.24) and may
+    need refinement if the phantom sits in a non-standard rotational orientation.
+    """
+
+    attr_name = "ctp515"
+    common_name = "Low Contrast (CTP730)"
+    roi_dist_mm = 55
+    roi_settings = {
+        # 1.0 % contrast group — upper-right to right (~2 o'clock)
+        "1pct_15mm": {"angle": -70, "distance": 55, "radius": 6.0},
+        "1pct_9mm":  {"angle": -60, "distance": 55, "radius": 3.5},
+        "1pct_8mm":  {"angle": -50, "distance": 55, "radius": 3.0},
+        "1pct_7mm":  {"angle": -40, "distance": 55, "radius": 2.5},
+        "1pct_6mm":  {"angle": -30, "distance": 55, "radius": 2.0},
+        "1pct_5mm":  {"angle": -20, "distance": 55, "radius": 1.5},
+        "1pct_4mm":  {"angle": -10, "distance": 55, "radius": 1.2},
+        "1pct_3mm":  {"angle":   0, "distance": 55, "radius": 1.0},
+        "1pct_2mm":  {"angle":  10, "distance": 55, "radius": 0.8},
+        # 0.5 % contrast group — upper-left to left (~10 o'clock)
+        "05pct_15mm": {"angle":  170, "distance": 55, "radius": 6.0},
+        "05pct_9mm":  {"angle":  180, "distance": 55, "radius": 3.5},
+        "05pct_8mm":  {"angle": -170, "distance": 55, "radius": 3.0},
+        "05pct_7mm":  {"angle": -160, "distance": 55, "radius": 2.5},
+        "05pct_6mm":  {"angle": -150, "distance": 55, "radius": 2.0},
+        "05pct_5mm":  {"angle": -140, "distance": 55, "radius": 1.5},
+        "05pct_4mm":  {"angle": -130, "distance": 55, "radius": 1.2},
+        "05pct_3mm":  {"angle": -120, "distance": 55, "radius": 1.0},
+        "05pct_2mm":  {"angle": -110, "distance": 55, "radius": 0.8},
+        # 0.3 % contrast group — bottom (~6 o'clock)
+        "03pct_15mm": {"angle":  50, "distance": 55, "radius": 6.0},
+        "03pct_9mm":  {"angle":  60, "distance": 55, "radius": 3.5},
+        "03pct_8mm":  {"angle":  70, "distance": 55, "radius": 3.0},
+        "03pct_7mm":  {"angle":  80, "distance": 55, "radius": 2.5},
+        "03pct_6mm":  {"angle":  90, "distance": 55, "radius": 2.0},
+        "03pct_5mm":  {"angle": 100, "distance": 55, "radius": 1.5},
+        "03pct_4mm":  {"angle": 110, "distance": 55, "radius": 1.2},
+        "03pct_3mm":  {"angle": 120, "distance": 55, "radius": 1.0},
+        "03pct_2mm":  {"angle": 130, "distance": 55, "radius": 0.8},
+    }
+    background_roi_dist_ratio = 0.75
+    background_roi_radius_mm = 4
+
+    @property
+    def rois_1pct(self) -> dict:
+        """ROIs belonging to the 1.0 % contrast group."""
+        return {k: v for k, v in self.rois.items() if k.startswith("1pct_")}
+
+    @property
+    def rois_05pct(self) -> dict:
+        """ROIs belonging to the 0.5 % contrast group."""
+        return {k: v for k, v in self.rois.items() if k.startswith("05pct_")}
+
+    @property
+    def rois_03pct(self) -> dict:
+        """ROIs belonging to the 0.3 % contrast group."""
+        return {k: v for k, v in self.rois.items() if k.startswith("03pct_")}
+
+
+# ---------------------------------------------------------------------------
+# CatPhan 604
+# ---------------------------------------------------------------------------
+
+@capture_warnings
+class CatPhan604(CatPhanBase):
+    """CatPhan 604.
+
+    Analyzes: HU linearity (CTP732), uniformity (CTP729), spatial resolution
+    via bead MTF (CTP528CP604), and low contrast (CTP730).
+    """
+
+    _demo_url = "CatPhan604.zip"
+    _model = "604"
+    catphan_radius_mm = 101
+    modules = {
+        CTP732: {"offset": 0},
+        CTP729: {"offset": -80},
+        CTP528CP604: {"offset": 0},   # bead co-located with HU plug slice
+        CTP730: {"offset": -40},
+    }
+
+    @staticmethod
+    def run_demo(show: bool = True):
+        cbct = CatPhan604.from_demo_images()
+        cbct.analyze()
+        print(cbct.results())
+        cbct.plot_analyzed_image(show)
+
+    def refine_origin_slice(self, initial_slice_num: int) -> int:
+        """Find the HU-linearity slice with the least wire-ramp angle.
+
+        Scans ±5 slices around ``initial_slice_num``, measures the horizontal
+        and vertical wire angles on each, and returns the slice with the
+        smallest mean angle.  This is robust to the RM R1-4 jig being present.
+        """
+        angles = []
+        ctp404, offset = self._get_module(CTP732, raise_empty=True)
+        original_setup = ctp404._setup_rois
+        ctp404._setup_rois = lambda x: x
+        ctp = ctp404(
+            self,
+            offset=offset,
+            clear_borders=self.clear_borders,
+            hu_tolerance=0,
+            scaling_tolerance=0,
+            thickness_tolerance=0,
+        )
+        ctp404._setup_rois = original_setup
+        for slice_num in range(initial_slice_num - 5, initial_slice_num + 5):
+            sl = Slice(self, slice_num, clear_borders=self.clear_borders)
+            troi = {}
+            for name, setting in ctp.thickness_roi_settings.items():
+                troi[name] = ThicknessROI.from_phantom_center(
+                    sl.image,
+                    setting["width_pixels"],
+                    setting["height_pixels"],
+                    setting["angle_corrected"],
+                    setting["distance_pixels"],
+                    sl.phan_center,
+                )
+            left_wire = troi["Left"].long_profile.center_idx
+            right_wire = troi["Right"].long_profile.center_idx
+            h_angle = abs(left_wire - right_wire)
+            top_wire = troi["Top"].long_profile.center_idx
+            bottom_wire = troi["Bottom"].long_profile.center_idx
+            v_angle = abs(top_wire - bottom_wire)
+            angle = (h_angle + v_angle) / 2
+            angles.append(
+                {
+                    "slice": slice_num,
+                    "angle": angle,
+                    "left width": troi["Left"].long_profile.field_width_px,
+                    "right width": troi["Right"].long_profile.field_width_px,
+                    "left center": troi["Left"].long_profile.y_at_x(left_wire),
+                    "right center": troi["Right"].long_profile.y_at_x(right_wire),
+                    "left profile": troi["Left"].long_profile.values,
+                    "right profile": troi["Right"].long_profile.values,
+                }
+            )
+
+        median_width_l = np.median([a["left width"] for a in angles])
+        median_width_r = np.median([a["right width"] for a in angles])
+        median_width = (median_width_l + median_width_r) / 2
+        median_left_pixel_val = np.median(
+            np.concatenate([a["left profile"] for a in angles])
+        )
+        median_right_pixel_val = np.median(
+            np.concatenate([a["right profile"] for a in angles])
+        )
+        median_pixel_val = (median_left_pixel_val + median_right_pixel_val) / 2
+        max_left_pixel_val = np.max(
+            np.concatenate([a["left profile"] for a in angles])
+        )
+        max_right_pixel_val = np.max(
+            np.concatenate([a["right profile"] for a in angles])
+        )
+        max_pixel_val = (max_left_pixel_val + max_right_pixel_val) / 2
+
+        for angle_set in angles.copy():
+            if (
+                angle_set["left width"] < median_width * 0.7
+                or angle_set["right width"] < median_width * 0.7
+            ):
+                angles.remove(angle_set)
+                continue
+            fwxm_pixel = np.mean(
+                (angle_set["left center"], angle_set["right center"])
+            )
+            delta_median = abs(median_pixel_val - fwxm_pixel)
+            delta_max = abs(max_pixel_val - fwxm_pixel)
+            if delta_median < delta_max:
+                angles.remove(angle_set)
+
+        m_slice_num = np.argsort([a["angle"] - self.catphan_roll for a in angles])
+        return angles[m_slice_num[0]]["slice"]
+
+
+# ---------------------------------------------------------------------------
+# Backward-compatibility aliases (old pylinac / upstream names)
+# ---------------------------------------------------------------------------
+
+#: Old name for :class:`CTP732`.
+CTP404CP604 = CTP732
+
+#: Old name for :class:`CTP729`.
+CTP486 = CTP729
+
+#: CTP730 dispatches via the CTP515 ABC; re-export the old name.
+CTP515 = CTP730
